modules/prompt: make steeef prompt use ${PRETTY_HOST} as hostname
By default the `steeef` prompt uses `%m` to print the full hostname,                 
such as `pdDDsEvGza.example.com`. When `ssh`ing between multiple hosts               
that are all using the `zprezto` with the `steeef` prompt, it can be                 
hard to remember which host is which, and we may not have permission                 
to change `${HOST}`/`{$HOSTNAME}` if the hosts are corporate machines           
or infrastructure.                                                              
                                                                               
This patch modifies the `steeef` prompt to use the `${PRETTY_HOST}`             
environment variable instead of `%m`. Users will need to export this            
in their `.zshrc` before `zprezto` is `source`d.